+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - ReSpecTh->ChemKED converter function now returns a dictionary, while the command-line entry points write out files
 - Require Habanero>=0.6.0 to support the `mailto` argument
 - Require pytest>=3.2.0 to support the `pytest.mark.filterwarnings` decorator
+- Deprecate the `volume-history` field in the ChemKED YAML file and replace with `time-histories`
 
 ### Fixed
 - Crossref lookups in the converters use the common API instance from validation